+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# 0.2.1
+# Changelog
+
+## 0.2.2
+
+ - Add item tooltip description support
+ - ItemBlock allows subtypes
+ - Add test module
+
+## 0.2.1
 
  - Add ability to disable optional module in config file
  - Fix formatting error with AdvancedFontRenderer